requirements.txt
numpy
pandas
scikit-learn
request
streamlit